+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - `Profile.ToModelCurves()`
+- `Profile.Difference()`
+- `Profile.Intersection()`
 - `Profile.Segments()`
 - `Bbox3.ToModelCurves()`
 - `Line.ExtendTo(IEnumerable<Line>)`